% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detect.R
\name{anomaly_detect}
\alias{anomaly_detect}
\title{Detect anomalies in time series}
\usage{
anomaly_detect(data = offers, x = "Datetime", y = "VWAP")
}
\arguments{
\item{data}{data.frame that should contain columns that can form a time series. defaults to inbuilt offers data set}

\item{x}{date format vector. defaults to 'Datetime' column in offers data set}

\item{y}{numeric vector of response data to find outlying points from. defaults to 'VWAP' column in offers data set}
}
\value{
anomaly class object with data on outlying points and plots of recomposed and decomposed data
}
\description{
Uses the anomalize package to run anomaly detection using the Seasonal Hybrid ESD algorithm, decomposing time 
series data into its components, plotting them and identifying outlying points in red.
}
\examples{
anomaly_detect(data = offers, x = 'Datetime', y = 'Megawatts', gsplit=FALSE)
}
