% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_points.R
\name{change_points}
\alias{change_points}
\title{Detect change points in time series}
\usage{
change_points(
  data = offers,
  x = "DollarsPerMegawattHour",
  group = "TradingDate"
)
}
\arguments{
\item{data}{data.frame that should contain columns that can form a time series. defaults to inbuilt offers data set}

\item{x}{numeric vector of response data to find outlying points from. defaults to 'DollarsPerMegawattHour' column in offers data set}

\item{group}{optional vector to group data by}
}
\value{
offers_changepoint class object with attribute 'changepoint' containing the indices of identified change points
}
\description{
Uses the changepoint.np package which implements the multiple changepoint algorithm PELT with a nonparametric cost function 
based on the empirical distribution of the data. This finds points delimiting segments where the mean and variance of a 
series has changed.
}
\examples{
change_points(data=offers,x='Megawatts', group='TradingDate')
plot(change_points())
}
