% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Visualization of correlation matrix}
\usage{
correlation(data = list(offers, demand))
}
\arguments{
\item{data}{a list of data.frames which should have at one column in common.}
}
\description{
Plots a correlation matrix, part numeric, part shape represented. Non-numeric columns are included.
}
\examples{
correlation(data = c(offers,demand))
}
