% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Generates html report}
\usage{
create_report(
  data = offers,
  response = "DollarsPerMegawattHour",
  report_title = "Report",
  output_dir = getwd(),
  output_file = "report.html"
)
}
\arguments{
\item{data}{data.frame that should contain columns that can form a time series. defaults to inbuilt offers data set}

\item{response}{numeric vector of the variable that should be the focus of the reports analysis, y argument in many functions. defaults to 'DollarsPerMegawattHour' from offers}

\item{report_title}{string of title of the report. defaults to 'Report'}

\item{output_dir}{get directory function or string of local path where report should be saved. defaults to current working directory}

\item{output_file}{string of html file name. defaults to 'report.html'}
}
\value{
does not return an object but generates an html file.
}
\description{
Generates an html file using the other functions in the package that should give the user analysis on the inputted 
datasets with summaries, plots and highlighting any anomalies in the data.
}
\examples{
create_report(data = offers, response = 'Megawatts')
create_report(data = offers, response = 'DollarsPerMegawattHour', report_title = "Report", output_dir = getwd(), output_file = "report.html")
}
