% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_offer_curve.R
\name{plot_offer_curve}
\alias{plot_offer_curve}
\title{Plot offer curve}
\usage{
plot_offer_curve(
  data = offers,
  x = "Megawatts",
  y = "DollarsPerMegawattHour",
  group = NULL
)
}
\arguments{
\item{data}{data.frame that should contain at least two numeric columns. defaults to inbuilt offers data set}

\item{x}{a numeric vector that should be able to be cumulatively summed by param y. defaults to 'Megawatts' from offers}

\item{y}{a numeric vector with matching param x values. defaults to 'DollarsPerMegawattHour' from offers}

\item{group}{a vector which the data can be grouped by. defaults to NULL}
}
\value{
a ggplot object plotting an offer curve
}
\description{
Plots offer curve with accumulated offers amounts, default megawatts, in the x axis and the values the offers, 
default dollars per megawatt hours, were made at in the y axis. Plots with the group option plots the offer curve 
for each group.
}
\examples{
plot_offer_curve(data = offers, x='Megawatts', y='DollarsPerMegawattHour', group='ParticipantCode')
plot_offer_curve(data = offers, x='DollarsPerMegawattHour', y='Megawatts', group='ParticipantCode')
}
